feat: fix navigation latency and implement workstation appearance themes & density
- Optimize API client with 127.0.0.1 default to eliminate Windows IPv6 2s connect delay
- Refactor SSE analytics into resilient shared singleton with exponential backoff
- Add instant navigation loading skeleton boundary in dashboard layout
- Implement global ThemeProvider supporting Cleanroom Day, Low-Glare Night, and System Sync
- Add complete obsidian dark mode and compact cleanroom density styles in globals.css
- Connect Settings workstation appearance controls to global theme context
- Add pre-hydration head script to prevent theme flashing on page reload
- Fix synthetic_images int() call warning and verify-demo script title matching

+++ backend/tests/fixtures/synthetic_images.py
@@ -50,7 +50,7 @@ def create_proportional_dot_image(
     fmt: str = ".png",
 ) -> bytes:
     """Create an image where dot radius is proportional to image size (e.g. size * normalized_radius)."""
-    radius = int(round(size * normalized_radius))
+    radius = round(size * normalized_radius)
     return create_centered_dot_image(size, radius, bg_color, dot_color, fmt)
